Fixed lingering positioning issues for the planar array. Tested with 5x2 and 3x3 and 30 channels each, looks like it puts them all in the right place now.

+++ Source/RxComponents/LMCAntennaElementPositioner.hh
@@ -33,7 +33,7 @@ namespace locust
             virtual bool Configure( const scarab::param_node& aNode );
             virtual double GetPositionZ(double zShiftArray, int channelIndex, int nChannels,
             		int nSubarrays, int nReceivers, double elementSpacingZ, int receiverIndex);
-            virtual double GetTheta(int channelIndex, int dThetaArray, int receiverIndex, int nReceivers, double elementRadius);
+            virtual double GetTheta(int channelIndex, double dThetaArray, int receiverIndex, int nReceivers, double elementRadius);
 			virtual void PlaceElement(Receiver &modelElement, double elementRadius, double theta, double zPosition);
             virtual bool IsPlanarArray() {return false;};
